Move engine's Web API calls from SV to Com
It makes more sense to have the Web API handled outside SV because there won't be a server running if a map isn't in-play. The Web API now starts in Com_Init and stops in Com_Shutdown, so it effectively lasts the whole process (similar to the Q3 event journaling system). Integration with libfcgi is also easier as it doesn't seem to like shutting down cleanly.

Also:
* Add safety guards to public-facing Web API functions to ensure it's in a valid state
* Close thread/event handles properly on shutdown