Implement shared FCGX_Request object
All the logic for handling the request is now in the main thread. Details of the request are printed to the console, and basic info (for testing purposes) about the game are returned to the web server.

WebAPI_Frame will now also shutdown the Web API if the accepting thread has exited.